Added license and expanded README
A LICENSE.md file has been added to the project, outlining permissions and restrictions for use of the software. The README.md file has been significantly expanded with a detailed description of the project, its purpose, contents of the repo, and limitations. An image was also added to visually represent some concepts explained in the README.